pyproject.toml
[project]
name = "cv_writer"
version = "0.1.0"
description = "CV Optimizer using crewAI Flow"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=1.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pypdf>=3.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "click>=8.1.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-ollama>=0.1.0",
]

[project.scripts]
cv-optimizer = "cv_writer.main:main"
plot = "cv_writer.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
